+++ src/gui/guiWire.cpp
@@ -439,7 +439,7 @@ void guiWire::setState(vector<StateType> state) {
 };
 
 void guiWire::setSubState(IDType buslineId, StateType state) {
-	for (int i = 0; i < this->state.size(); i++) {
+	for (int i = 0; i < (int)this->state.size(); i++) {
 		if (ids[i] == buslineId) {
 			this->state[i] = state;
 		}
